+++ spec/serializers/game_deserializer_spec.rb
@@ -64,4 +64,25 @@ describe GameDeserializer do
       end
     end
   end
+
+  context 'after cities are built' do
+    before(:each) do
+      game.handle(@setup_game_interactor)
+      game.handle(GrantCity.new(spot: 1, player: game.players[0]))
+      game.handle(GrantCity.new(spot: 10, player: game.players[1]))
+    end
+
+    it 'deserializes to equal game object' do
+      expect(call).to eq(game)
+    end
+
+    context 'when serialized to json' do
+      let(:json) { serializer.call.to_json }
+      let(:hash) { JSON.parse(json) }
+
+      it 'deserializes from json' do
+        expect(call).to eq(game)
+      end
+    end
+  end
 end